Renamed client.tests.js to oauth.tests.js
Access tokens tested with appropriate lichess API.
Tests are more descriptive.
